% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAng.R
\name{calcAng}
\alias{calcAng}
\title{Calculate POSITIVE angles}
\usage{
calcAng(y, x, degrees = FALSE)
}
\arguments{
\item{y}{A vector y-coordinates}

\item{x}{A vector of x-coordinates}

\item{degrees}{If TRUE angles are given in degrees, otherwise (default) they
are given in radians}
}
\value{
A vector equal length to x and y of angles. x and y coordinates are
used by pairs (i.e. the first vector has coordinates c(x\link{1}, y\link{1}), the
second c(x\link{2}, y\link{2}), etc)/
}
\description{
Calculates an angle between \link{0, 2*pi} from the x-axis to the
vector that goes from the origin to the point x, y
}
\examples{
x <- seq(-3, 3, 1); y = seq(-3, 3, 1)
calcAng(y, x)
}
