% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDist.R
\name{calcDist}
\alias{calcDist}
\title{Calculate Euclidean distance}
\usage{
calcDist(s, x, y)
}
\arguments{
\item{s}{A numeric vector with two coordinates c(x, y), in this order, that will be used as the point to calculate distance from.}

\item{x}{A numeric  vector of x-coordinates of the points to calculate distance to. Same length as y.}

\item{y}{A numeric vector of y-coordinates of the points to calculate distance to. Same length as x.}
}
\value{
A vector of distances as long as the length of x and y.
}
\description{
Helper function to vectorize distance calculations over multiple points
}
\examples{
calcDist(c(0, 0), c(1, 2, 3), c(1, 2, 3))
}
