% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUdColony.R
\name{calcUdCol}
\alias{calcUdCol}
\title{Calculate utilization distribution around a colony}
\usage{
calcUdCol(.col_sel, .age, .hab = NULL, .scale = FALSE, .countsdir, .outputdir)
}
\arguments{
\item{.col_sel}{A named vector (data frame row) with at least three elements:
c("lon", "lat", "id", "avg_ad", "avg_juv")}

\item{.age}{Age of the vultures a utilization distribution is calculated for.}

\item{.hab}{A dataframe with the longitude ("lon") latitude ("lat") of the
grid cells the UD is computed at. If NILL, the object "range_covts" will be
used.}

\item{.scale}{TRUE/FALSE whether the UD should be scaled by the number of
individuals expected to use the colony.}

\item{.countsdir}{A character string with the directory where the smoothed
counts are stored.}

\item{.outputdir}{A character string with the directory to store the results.}
}
\value{
A dataframe with the counts and smoothed counts computed for each
grid cell in .hab.
}
\description{
Calculate utilization distribution around a colony
}
