% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUdHgtColony.R
\name{calcUdHgtColony}
\alias{calcUdHgtColony}
\title{Calculate utilization distribution at rotor height around a colony}
\usage{
calcUdHgtColony(
  .col_sel,
  .age,
  .hab = NULL,
  .scale = FALSE,
  .countsdir,
  .suffix,
  .outputdir = NULL
)
}
\arguments{
\item{.col_sel}{A named vector (data frame row) with at least three elements:
c("lon", "lat", "id", "avg_ad", "avg_juv")}

\item{.age}{Age of the vultures a utilization distribution is calculated for.}

\item{.hab}{A dataframe with the longitude ("lon") latitude ("lat") of the
grid cells the UD is computed at. If NULL, the object "range_covts" will be
used.}

\item{.scale}{TRUE/FALSE whether the UD should be scaled by the number of
individuals expected to use the colony.}

\item{.countsdir}{A character string with the directory where the smoothed
counts(expected counts, in this case) are stored.}

\item{.suffix}{A character string containing the part the name of the file
that appears after the colony code (e.g. "_ad_hgt_gam.rds)}

\item{.outputdir}{A character string with the directory to store the results.
If NULL (default) the output is not saved to disk.}
}
\value{
A dataframe with the expected counts and smoothed expected counts
computed for each grid cell in .hab.
}
\description{
Calculate utilization distribution at rotor height around a colony
}
