% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUDquantile.R
\name{calcUDquantile}
\alias{calcUDquantile}
\title{Calculate utilization distribution quantile}
\usage{
calcUDquantile(ud, probs)
}
\arguments{
\item{ud}{A vector of utilization intensity values}

\item{probs}{A vector of proportions/probabilities.}
}
\value{
A vector with the same length of probs, representing the utilization
intensity values that capture the desired proportion of activity/probability
of occurrence.
}
\description{
Calculates the CDF of a utilization distribution, passed on as
a vector of densities/probabilities. It returns the highest value that
captures the desired level of activity proportion.
}
