% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeDataFrame.R
\name{completeDataFrame}
\alias{completeDataFrame}
\title{Complete data frame}
\usage{
completeDataFrame(df, allvars, value)
}
\arguments{
\item{df}{Data frame with missing variables}

\item{allvars}{A character vector with the names of all the variables the data frame should have}

\item{value}{The value that should have the missing variables by default}
}
\value{
A data frame with all the variables contained in 'df' and 'allvars'. Those variables originally contained in 'allvars' and not in 'df' will be added to 'df' and given a value
specified in the 'value' field. One would tipically fill in those variable with some desired values.
}
\description{
Takes a data frame and adds a number of missing variables filling them with NA or other value provided
}
\examples{
completeDataFrame(range_covts, c("disturbed", "dist_col:ttnoon"), 0)
}
