% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estHgtRisk.R
\name{estHgtRisk}
\alias{estHgtRisk}
\title{Estimate height risk}
\usage{
estHgtRisk(.col_sel, .coefs, .datadir, .simsdir, .outdir)
}
\arguments{
\item{.col_sel}{A named vector with the coordinates of the colony we want to
simulate around c("lon", "lat").}

\item{.coefs}{A list with samples of random coefficients. Each element in the
list corresponds to a set of coefficients for a different individual.}

\item{.simsdir}{A character string with the path to the directory containing
the points for which height risk wants to be estimated.}

\item{.outdir}{A character string with the path to the directory where the
results are to be stored at in disk. If NULL the results are not stored.}

\item{data_dir}{A character string with the path to the directory containing
the auxiliary data needed for the estimation (i.e. colony and supplementary
feeding sites data).}
}
\value{

}
\description{
Estimate height risk
}
