% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgt_fit_summary.R
\docType{data}
\name{hgt_fit_summary}
\alias{hgt_fit_summary}
\title{Summary of the flight height model fitted to the Cape Vulture data}
\format{
A list with the output of the glmmTMB::summary function.
We are most interested in the regression coefficients:
\describe{
\item{\code{item}}{Description}
}
}
\usage{
hgt_fit_summary
}
\description{
Summary of the flight height model fitted to the Cape Vulture data
}
\details{
DETAILS
}
\keyword{datasets}
