% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make2DSEmap.R
\name{make2DSEmap}
\alias{make2DSEmap}
\title{Make 2D standard error map}
\usage{
make2DSEmap(col_to_pred, age, sedir, suffix, outputdir = NULL)
}
\arguments{
\item{col_to_pred}{A dataframe with the coordinates of all colonies we want
to include in the map.}

\item{age}{A character string indicating the age of the vultures we want to
calculate the risk map for: either "ad" or "juv.}

\item{sedir}{A character string indicating the path to the directory
where the SE files are.}

\item{suffix}{A character string containing the part the name of the file
that appears after the colony code (e.g. "_ssf_se.rds)}

\item{outdir}{A character string. If not NULL (default) the output map is
saved to disk in the desired location. The map type and the age will be
recorded in the name of the output file.}
}
\value{

}
\description{
Takes all files with step-selection standard error present in a
directory and creates a SE map.
}
