% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make3DRiskMap.R
\name{make3DRiskMap}
\alias{make3DRiskMap}
\title{Make 3D UD Map}
\usage{
make3DRiskMap(
  col_to_pred,
  age,
  map_type = "pud",
  countsdir,
  suffix,
  outdir = NULL
)
}
\arguments{
\item{col_to_pred}{A dataframe with the coordinates and sizes of all colonies
we want to include in the map.}

\item{age}{A character string indicating the age of the vultures we want to
calculate the UD map for: either "ad" or "juv.}

\item{map_type}{A character string indicating the type of map that should be
calculated: either "pud" in which the utilization distribution is multiplied
by the size of the colony or "ud" in which the UD is not scaled by the
size of the colony.}

\item{countsdir}{A character string indicating the path to the directory
where the smoothed counts are found.}

\item{outdir}{A character string. If not NULL (default) the output map is
saved to disk in the desired location. The map type and the age will be
recorded in the name of the output file.}

\item{.suffix}{A character string containing the part the name of the file
that appears after the colony code (e.g. "_ad_hgt_gam.rds)}
}
\description{
Takes all files with smoothed counts present in a directory
and creates a UD map.
}
