% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_codes.R
\docType{data}
\name{map_codes}
\alias{map_codes}
\title{Map codes}
\format{
A character vector
}
\usage{
map_codes
}
\description{
These are the codes of the maps used for fitting the SSF model
and that define most of the distribution range of the Cape Vulture
}
\details{
DETAILS
}
\keyword{datasets}
