% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov_kernel.R
\docType{data}
\name{mov_kernel}
\alias{mov_kernel}
\title{Movement kernel}
\format{
A named vector with three elements:
\describe{
\item{\code{shape}}{The shape parameter of the gamma distribution}
\item{\code{scale}}{The scale parameter of the gamma distribution}
\item{\code{model_sc}}{The factor used to scale step-length by to fit the step-selection model}
}
}
\usage{
mov_kernel
}
\description{
Parameters of the gamma density fitted to the step-lengths of the vultures with one hour-resolution tracking data. This is the preliminary distribution fitted before fitting the linear model.
}
\details{
DETAILS
}
\keyword{datasets}
