% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinom_dur.R
\docType{data}
\name{nbinom_dur}
\alias{nbinom_dur}
\title{Negative binomial trip-duration distribution}
\format{
A list with two elements, one for adults and one for juveniles, each containing:
\describe{
\item{\code{size}}{The size parameter of the negative binomial distribution}
\item{\code{mu}}{The mu parameter of the negative binomial distribution}
}
}
\usage{
nbinom_dur
}
\description{
Parameters of the negative binomial distribution that describes the duration of the trips vultures make away from the colony
}
\details{
DETAILS
}
\keyword{datasets}
