% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepColHab.R
\name{prepColHab}
\alias{prepColHab}
\title{Prepare habitat around a colony}
\usage{
prepColHab(col_cc, max_range, col_all = NULL, sfs = NULL)
}
\arguments{
\item{col_cc}{A numeric vector with the coordinates of the colony to simulate around (lon, lat)}

\item{max_range}{The maximum distance from the colony the vultures will be allowed to travel in kilometers.
Note that increasing the range will increase computation time and memory use}
}
\value{
A data frame with variables needed for activity simulation such as: distance to colony, distance to other colonies
or distance to supplementary feeding sites.
}
\description{
Prepare habitat around a colony
}
\examples{
prepColHab(c(27.7, -31.2), 100)
}
