% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_covts.R
\docType{data}
\name{range_covts}
\alias{range_covts}
\title{range_covts}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{\code{first}}{character COLUMN_DESCRIPTION}
\item{\code{second}}{double COLUMN_DESCRIPTION}
\item{\code{third}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
range_covts
}
\description{
General covariates across the Cape Vulture range. These covariates need to be adapted for specific colonies
}
\details{
DETAILS
}
\keyword{datasets}
