% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleHgtCoef.R
\name{sampleHgtCoef}
\alias{sampleHgtCoef}
\title{Sample height flight model coefficients}
\usage{
sampleHgtCoef(mod_summary = hgt_fit_summary, nind = 1, seed = NULL)
}
\arguments{
\item{mod_summary}{The summary of a glmmTMB model. If not specified,
the hgt_fit_summary, contained in the package data will be used}

\item{nind}{Number of individuals to sample for}

\item{seed}{Whether a seed should be specified to the random selection of
coefficients}
}
\value{
A list with each element being a vector of step-selectioncoefficients
}
\description{
Sample random height flight model coefficients from a
hierarchical height flight model.
}
\examples{
vultRmap::sampleHgtCoef(nind = 2)
}
