% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSsfCoef.R
\name{sampleSsfCoef}
\alias{sampleSsfCoef}
\title{Sample step-selection coefficients}
\usage{
sampleSsfCoef(mod_summary = ssf_fit_summary, nind = 1, seed = NULL)
}
\arguments{
\item{mod_summary}{The summary of a glmmTMB model. If not specified, the ssf_fit_summary, contained in the package data will be used}

\item{nind}{Number of individuals to sample for}

\item{seed}{Whether a seed should be specified to the random selection of coefficients}
}
\value{
A list with each element being a vector of step-selection coefficients
}
\description{
Sample random step-selection coefficients from a hierarchical step-selection model
}
\examples{
vultRmap::sampleSsfCoef(nind = 2)
}
