% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_height_risk.R
\name{sim_height_risk}
\alias{sim_height_risk}
\title{Simulate risk height from step-selection activity simulations}
\usage{
sim_height_risk(colonies = "all", data_dir, sims_dir, out_dir, ncoefs, seed)
}
\arguments{
\item{colonies}{Either a vector of colony codes to simulate activity around
or "all" to simulate activity around all colonies that meet certain criteria.
At the moment these are breeding colonies for which there are counts and roots
for which there are counts of more than 50 birds.}

\item{data_dir}{Directory where the colony data is.}

\item{sims_dir}{Directory where the step-selection activity simulations are stored}

\item{out_dir}{Directory where the output of the simulations should go to}

\item{ncoefs}{Number of random coefficients to be sampled from the risk height
model.}

\item{seed}{A random seed that will be used to sample random model coefficients}
}
\value{

}
\description{
Simulate risk height from step-selection activity simulations
}
