% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ssf.R
\name{sim_ssf}
\alias{sim_ssf}
\title{Simulate activity from step-selection}
\usage{
sim_ssf(colonies = "all", out_dir, data_dir, seeds, ...)
}
\arguments{
\item{colonies}{Either a vector of colony codes to simulate activity around
or "all" to simulate activity around all colonies that meet certain criteria.
At the moment these are breeding colonies for which there are counts and roots
for which there are counts of more than 50 birds.}

\item{out_dir}{Directory where the output of the simulations should go to}

\item{data_dir}{Directory where the colony data is.}

\item{seeds}{A vector of two seeds that will be used in the simulations}

\item{...}{Other parameters to be passed on to \href{simAllColonies}{vultRmap}}
}
\value{

}
\description{
Simulate activity from step-selection
}
