% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAllColonies.R
\name{simAllColonies}
\alias{simAllColonies}
\title{Simulate activity for all colonies}
\usage{
simAllColonies(
  age,
  min_size_col = 1,
  min_size_roost = 50,
  out_dir,
  data_dir,
  ...
)
}
\arguments{
\item{age}{Age of the vultures: "ad" for adult, "juv" for juveniles.}

\item{min_size_col}{Simulations will only be produced for colonies with number of
adults greater or equal to min_size_col.}

\item{min_size_roost}{Simulations will only be produced for roosts with number of
birds (adults + juveniles) greater or equal to min_size_roost.}

\item{out_dir}{Specifies the directory where the simulation results should
be saved to.}

\item{data_dir}{A character string with the path to the directory containing
the auxiliary data needed for the simulations.}

\item{...}{Other arguments passed on to simOneColony.}
}
\value{
A data frame with at least the desired number of steps (the same
number of steps are allocated to cores and rounded up) for each colony. Due
to potentially large output an output directory must be specified where a data
frame for each colony will be saved.
}
\description{
This is a wrapper around simOneColony that does all necessary
computation for simulating activity around all colonies.
}
