% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOneColony.R
\name{simOneColony}
\alias{simOneColony}
\title{Simulate activity for one colony.}
\usage{
simOneColony(
  col_sel,
  totalsteps,
  dist_lim,
  age,
  sample_coefs = NULL,
  set_seed = NULL,
  ncores = 1,
  data_dir = "../vultRmap_data_aux"
)
}
\arguments{
\item{col_sel}{A named vector with the coordinates of the colony we want to
simulate around c("lon", "lat").}

\item{totalsteps}{Total number of steps we want to simulate}

\item{dist_lim}{Maximum distance vultures are allowed to travel away from
its central colony (col_sel).}

\item{age}{Age of the vultures: "ad" for adult, "juv" for juveniles}

\item{sample_coefs}{Whether step-selection coefficients should be sampled
from the distribution of random effects. Either NULL in which case the mean
of the distribution of effects will be used, or a number specifying the
number of coefficients to sample.}

\item{set_seed}{Specifies whether a random seed should be specified. Either
NULL for no random seed or a number specifying the seed to use.}

\item{ncores}{Specifies the number of cores that should be used for the
simulations.}

\item{data_dir}{A character string with the path to the directory containing
the auxiliary data needed for the simulations.}
}
\value{
A data frame with at least the desired number of steps (the same
number of steps are allocated to cores and rounded up).
}
\description{
This is a wrapper around simTrips that does all necessary
computation for simulating activity around a colony.
}
