% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrips.R
\name{simTrips}
\alias{simTrips}
\title{Simulate trips from colony}
\usage{
simTrips(.nsteps, .age, .hab, .mov_ker, .ssf_coef, .col_sel)
}
\arguments{
\item{.nsteps}{Total number of steps to simulate}

\item{.age}{A character string indicating the age of the vulture to simulate activity for. Either 'ad' or 'juv'.}

\item{.hab}{Covariate values of the grid of cells representing the space the vultures can move through.
Each row of the data frame must represent one grid cell and must have, at least the coordinates in latitude/longitude
and the projected coordinates (needed for step-length calculations).}

\item{.mov_ker}{A named vector of parameters containing: i) the 'shape' and ii) the 'scale' of the gamma distribution
of step-lengths, and iii) 'model_sc', the scaling factor used for step-length when fitting the model.}

\item{.ssf_coef}{A vector of step-selection coefficients.}

\item{.col_sel}{A named vector with the coordinates (lon,lat) of the colony we want to simulate for.}
}
\value{
A data frame with the location of steps (lon, lat), distance from the central colony,
time of day (time-to-noon), trip number and duration of the trip (time away from colony).
A bird is considered to be away from the colony when it is further than 10km.
}
\description{
Simulate trips from colony
}
