% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssf_fit_summary.R
\docType{data}
\name{ssf_fit_summary}
\alias{ssf_fit_summary}
\title{Summary of the step-selection function model fitted to the Cape Vulture data}
\format{
A list with the output of the glmmTMB::summary function. We are most interested in the regression coefficients:
\describe{
\item{\code{item}}{Description}
}
}
\usage{
ssf_fit_summary
}
\description{
Summary of the step-selection function model fitted to the Cape Vulture data
}
\details{
DETAILS
}
\keyword{datasets}
