% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_colonies.R
\docType{data}
\name{test_colonies}
\alias{test_colonies}
\title{Test colonies}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{\code{id}}{Character. Identification code}
\item{\code{lon}}{Double. Longitude in decimal degrees}
\item{\code{lat}}{Double. Latitude in decimal degrees}
}
}
\usage{
test_colonies
}
\description{
Five colonies selected for testing the simulation algorithms
}
\details{
DETAILS
}
\keyword{datasets}
