% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sup_feeding.R
\docType{data}
\name{test_sup_feeding}
\alias{test_sup_feeding}
\title{Test supplementary feeding sites}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{\code{id}}{Character. Identification code}
\item{\code{lon}}{Double. Longitude in decimal degrees}
\item{\code{lat}}{Double. Latitude in decimal degrees}
}
}
\usage{
test_sup_feeding
}
\description{
Five supplementary feeding sites selected for testing the simulation algorithms
}
\details{
DETAILS
}
\keyword{datasets}
