% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateColonyCount.R
\name{updateColonyCount}
\alias{updateColonyCount}
\title{Update colony information}
\usage{
updateColonyCount(
  id_colony = NULL,
  ad_new = NULL,
  prop_juv = 0.43,
  remove_col = FALSE,
  add_col = FALSE,
  info_col = NULL,
  dir_colony,
  overwrite
)
}
\arguments{
\item{id_colony}{A character string corresponding to the colony that needs to
be updated.}

\item{ad_new}{The updated number of adults the site.}

\item{prop_juv}{Proportion of juveniles to adults. By default this is 0.43
which is a rough estimate based on the literature.}

\item{remove_col}{If TRUE the colony with id_colony will be completely removed.
If a colony should act as an attractor of vulture activity or might host
vultures in the future it should NOT be removed, but allocated ad_new = NA.
Unless we know that a colony doesn't exist, we don't want to remove them.}

\item{add_col}{If TRUE a new record will be added to the database instead of
updated. Additional fields must be passed on to info_col.}

\item{info_col}{A named list with the following fields: "name", "lon", "lat",
"avg_ad", "avg_juv", "counted" (1 if it has been counted, 0 if not), "type"
(either "breed" or "roost"), "id" new id (if NULL an id will be assigned
automatically).}

\item{dir_colony}{The directory where the colony data is stored. The name and
extention of the file must also be supplied.}

\item{overwrite}{If TRUE the original file will be overwriten! Be sure to
make a copy just in case. If FALSE the new dataframe will not be saved to
disk.}
}
\value{
A colony dataframe with the updated colony count.
}
\description{
This function will update the colony database and potentially
overwrite the existing one. Make sure to make a backup copy before
proceeding with any updates!
}
