% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHabMinDist.R
\name{calcHabMinDist}
\alias{calcHabMinDist}
\title{Calculate minimum distances to habitat cells}
\usage{
calcHabMinDist(hab, origin, features, buffer = NULL)
}
\arguments{
\item{hab}{A dataframe with at least two columns containing the coordinates (lon, lat) of a set of locations. Potentially a lattice of habitat cells.}

\item{origin}{A vector with the two coordinates (lon, lat) of the central point that will be used:
i) to define the Mercator projected coordinate system used for distance calculations and
ii) we might want to exclude features that are close to this point (e.g. colonies).}

\item{features}{A dataframe with at least two columns containg the coordinates (lon, lat) of the locations we want to calculate minimum distances to.}

\item{buffer}{A number indicating a distance from origin. If set, features within this distance to origin will be ignored.}
}
\value{
A vector with the minimum distance from each of the hab locations to any of the feature locations.
}
\description{
Calculate minimum distances to habitat cells
}
\examples{
hab <- expand.grid(-10:10, -10:10)
names(hab) <- c("lon", "lat")
origin <- c(2, 1)
features <- data.frame(lon = c(-5, 5),
                       lat = c(-5, 5))
calcHabMinDist(hab, origin, features)
}
