% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRegionRisk.R
\name{calcRegionRisk}
\alias{calcRegionRisk}
\title{Calculate region risk}
\usage{
calcRegionRisk(region, ref_area = NULL, risk_map)
}
\arguments{
\item{region}{An \link[sf]{sf} polygon or multipolygon object to predict risk
for. Risk is given as number of individuals at risk.}

\item{ref_area}{Optional. An \link[sf]{sf} polygon object. if provided risk
is calculated relative to this reference area as a the proportion of
individuals at risk.}

\item{risk_map}{A \link[raster]{raster} object with from which risk values
are extracted from.}
}
\value{

}
\description{
Calculate region risk
}
