% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinom_dur.R, R/nbinom_dur10.R, R/nbinom_dur2.R
\docType{data}
\name{nbinom_dur}
\alias{nbinom_dur}
\title{Negative binomial trip-duration distribution 5 km}
\format{
A list with two elements, one for adults and one for juveniles, each containing:
\describe{
\item{\code{size}}{The size parameter of the negative binomial distribution}
\item{\code{mu}}{The mu parameter of the negative binomial distribution}
}

A list with two elements, one for adults and one for juveniles, each containing:
\describe{
\item{\code{size}}{The size parameter of the negative binomial distribution}
\item{\code{mu}}{The mu parameter of the negative binomial distribution}
}

A list with two elements, one for adults and one for juveniles, each containing:
\describe{
\item{\code{size}}{The size parameter of the negative binomial distribution}
\item{\code{mu}}{The mu parameter of the negative binomial distribution}
}
}
\usage{
nbinom_dur

nbinom_dur

nbinom_dur
}
\description{
Parameters of the negative binomial distribution that describes the duration of the trips vultures make away from the colony
considering that a trip starts when the vulture is 5 km away from its central colony

Parameters of the negative binomial distribution that describes the duration of the trips vultures make away from the colony
considering that a trip starts when the vulture is 10 km away from its central colony

Parameters of the negative binomial distribution that describes the duration of the trips vultures make away from the colony
considering that a trip starts when the vulture is 2 km away from its central colony
}
\details{
DETAILS

DETAILS

DETAILS
}
\keyword{datasets}
