% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAllColonies.R
\name{simAllColonies}
\alias{simAllColonies}
\title{Simulate activity for all colonies}
\usage{
simAllColonies(
  totalsteps,
  dist_lim,
  age,
  sample_coefs = NULL,
  set_seed = NULL,
  ncores = 1,
  out_dir
)
}
\arguments{
\item{totalsteps}{Total number of steps we want to simulate for each colony}

\item{dist_lim}{Maximum distance vultures are allowed to travel away from
its central colony.}

\item{age}{Age of the vultures: "ad" for adult, "juv" for juveniles}

\item{sample_coefs}{Whether step-selection coefficients should be sampled
from the distribution of random effects. Either NULL in which case the mean
of the distribution of effects will be used, or a number specifying the
number of coefficients to sample.}

\item{set_seed}{Specifies whether a random seed should be specified. Either
NULL for no random seed or a number specifying the seed to use.}

\item{ncores}{Specifies the number of cores that should be used for the
simulations.}

\item{out_dir}{Specifies the directory where the simulation results should
be saved to.}
}
\value{
A data frame with at least the desired number of steps (the same
number of steps are allocated to cores and rounded up) for each colony. Due
to potentially large output an output directory must be specified where a data
frame for each colony will be saved.
}
\description{
This is a wrapper around simTrips that does all necessary
computation for simulating activity around all colony.
}
